Renamed ui package to rest package as it's more appropriate.
